refactor bisect
